+++ src/Credentials/ImpersonatedServiceAccountCredentials.php
@@ -30,6 +30,15 @@ use GuzzleHttp\Psr7\Request;
 use InvalidArgumentException;
 use LogicException;
 
+/**
+ * **IMPORTANT**:
+ * This class does not validate the credential configuration. A security
+ * risk occurs when a credential configuration configured with malicious urls
+ * is used.
+ * When the credential configuration is accepted from an
+ * untrusted source, you should validate it before creating this class.
+ * @see https://cloud.google.com/docs/authentication/external/externally-sourced-credentials
+ */
 class ImpersonatedServiceAccountCredentials extends CredentialsLoader implements
     SignBlobInterface,
     GetUniverseDomainInterface
@@ -78,11 +87,14 @@ class ImpersonatedServiceAccountCredentials extends CredentialsLoader implements
      *     @type string[]                       $delegates The delegates to impersonate
      * }
      * @param string|null $targetAudience The audience to request an ID token.
+     * @param string|string[]|null $defaultScope The scopes to be used if no "scopes" field exists
+     *                                           in the `$jsonKey`.
      */
     public function __construct(
         string|array|null $scope,
         string|array $jsonKey,
-        private ?string $targetAudience = null
+        private ?string $targetAudience = null,
+        string|array|null $defaultScope = null,
     ) {
         if (is_string($jsonKey)) {
             if (!file_exists($jsonKey)) {
@@ -101,6 +113,9 @@ class ImpersonatedServiceAccountCredentials extends CredentialsLoader implements
         if (!array_key_exists('source_credentials', $jsonKey)) {
             throw new LogicException('json key is missing the source_credentials field');
         }
+
+        $jsonKeyScope = $jsonKey['scopes'] ?? null;
+        $scope = $scope ?: $jsonKeyScope ?: $defaultScope;
         if ($scope && $targetAudience) {
             throw new InvalidArgumentException(
                 'Scope and targetAudience cannot both be supplied'
@@ -118,7 +133,13 @@ class ImpersonatedServiceAccountCredentials extends CredentialsLoader implements
                 // an ID token, the narrowest scope we can request is `iam`.
                 $scope = self::IAM_SCOPE;
             }
-            $jsonKey['source_credentials'] = CredentialsLoader::makeCredentials($scope, $jsonKey['source_credentials']);
+            $jsonKey['source_credentials'] = match ($jsonKey['source_credentials']['type'] ?? null) {
+                // Do not pass $defaultScope to ServiceAccountCredentials
+                'service_account' => new ServiceAccountCredentials($scope, $jsonKey['source_credentials']),
+                'authorized_user' => new UserRefreshCredentials($scope, $jsonKey['source_credentials']),
+                'external_account' => new ExternalAccountCredentials($scope, $jsonKey['source_credentials']),
+                default => throw new \InvalidArgumentException('invalid value in the type field'),
+            };
         }
 
         $this->targetScope = $scope ?? [];