+++ tests/ApplicationDefaultCredentialsTest.php
@@ -70,7 +70,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testLoadsDefaultFileIfPresentAndEnvVarIsNotSet()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures');
+        setHomeEnv(__DIR__ . '/fixtures');
         $this->assertNotNull(
             ApplicationDefaultCredentials::getCredentials('a scope')
         );
@@ -80,7 +80,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $this->expectException(DomainException::class);
 
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
         // simulate not being GCE and retry attempts by returning multiple 500s
         $httpHandler = getHandler([
             new Response(500),
@@ -93,7 +93,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testSuccedsIfNoDefaultFilesButIsOnGCE()
     {
-        putenv('HOME');
+        setHomeEnv(null);
 
         $wantedTokens = [
             'access_token' => '1/abdef1234567890',
@@ -116,7 +116,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testGceCredentials()
     {
-        putenv('HOME');
+        setHomeEnv(null);
 
         $jsonTokens = json_encode(['access_token' => 'abc']);
 
@@ -160,7 +160,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testImpersonatedServiceAccountCredentials()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures5');
+        setHomeEnv(__DIR__ . '/fixtures5');
         $creds = ApplicationDefaultCredentials::getCredentials(
             null,
             null,
@@ -183,7 +183,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testUserRefreshCredentials()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures2');
+        setHomeEnv(__DIR__ . '/fixtures2');
 
         $creds = ApplicationDefaultCredentials::getCredentials(
             null, // $scope
@@ -219,7 +219,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testServiceAccountCredentials()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures');
+        setHomeEnv(__DIR__ . '/fixtures');
 
         $creds = ApplicationDefaultCredentials::getCredentials(
             null, // $scope
@@ -255,7 +255,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testDefaultScopeArray()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures2');
+        setHomeEnv(__DIR__ . '/fixtures2');
 
         $creds = ApplicationDefaultCredentials::getCredentials(
             null, // $scope
@@ -292,7 +292,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testLGetMiddlewareoadsDefaultFileIfPresentAndEnvVarIsNotSet()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures');
+        setHomeEnv(__DIR__ . '/fixtures');
         $this->assertNotNull(ApplicationDefaultCredentials::getMiddleware('a scope'));
     }
 
@@ -300,7 +300,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $this->expectException(DomainException::class);
 
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
 
         // simulate not being GCE and retry attempts by returning multiple 500s
         $httpHandler = getHandler([
@@ -356,7 +356,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $this->expectException(DomainException::class);
 
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
 
         $mockCacheItem = $this->prophesize('Psr\Cache\CacheItemInterface');
         $mockCacheItem->isHit()
@@ -380,7 +380,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testOnGceCacheWithoutHit()
     {
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
 
         $gceIsCalled = false;
         $dummyHandler = function ($request) use (&$gceIsCalled) {
@@ -416,7 +416,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testOnGceCacheWithOptions()
     {
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
 
         $prefix = 'test_prefix_';
         $lifetime = '70707';
@@ -473,7 +473,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testGetIdTokenCredentialsLoadsDefaultFileIfPresentAndEnvVarIsNotSet()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures');
+        setHomeEnv(__DIR__ . '/fixtures');
         $creds = ApplicationDefaultCredentials::getIdTokenCredentials($this->targetAudience);
         $this->assertInstanceOf(ServiceAccountCredentials::class, $creds);
     }
@@ -483,7 +483,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
         $this->expectException(DomainException::class);
         $this->expectExceptionMessage('Your default credentials were not found');
 
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
 
         // simulate not being GCE and retry attempts by returning multiple 500s
         $httpHandler = getHandler([
@@ -500,7 +500,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testGetIdTokenCredentialsWithImpersonatedServiceAccountCredentials()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures5');
+        setHomeEnv(__DIR__ . '/fixtures5');
         $creds = ApplicationDefaultCredentials::getIdTokenCredentials('123@456.com');
         $this->assertInstanceOf(ImpersonatedServiceAccountCredentials::class, $creds);
     }
@@ -529,7 +529,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testGetIdTokenCredentialsSuccedsIfNoDefaultFilesButIsOnGCE()
     {
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
         $wantedTokens = [
             'access_token' => '1/abdef1234567890',
             'expires_in' => '57',
@@ -553,7 +553,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testGetIdTokenCredentialsWithUserRefreshCredentials()
     {
-        putenv('HOME=' . __DIR__ . '/fixtures2');
+        setHomeEnv(__DIR__ . '/fixtures2');
 
         $creds = ApplicationDefaultCredentials::getIdTokenCredentials(
             $this->targetAudience,
@@ -610,7 +610,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $quotaProject = 'quota-project-from-env-var';
         putenv(CredentialsLoader::QUOTA_PROJECT_ENV_VAR . '=' . $quotaProject);
-        putenv('HOME=' . __DIR__ . '/fixtures');
+        setHomeEnv(__DIR__ . '/fixtures');
 
         $credentials = ApplicationDefaultCredentials::getCredentials();
 
@@ -625,7 +625,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $quotaProject = 'quota-project-from-parameter';
         putenv(CredentialsLoader::QUOTA_PROJECT_ENV_VAR . '=quota-project-from-env-var');
-        putenv('HOME=' . __DIR__ . '/fixtures');
+        setHomeEnv(__DIR__ . '/fixtures');
 
         $credentials = ApplicationDefaultCredentials::getCredentials(
             null, // $scope
@@ -688,7 +688,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
 
     public function testWithGCECredentials()
     {
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
         $wantedTokens = [
             'access_token' => '1/abdef1234567890',
             'expires_in' => '57',
@@ -721,7 +721,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     public function testAppEngineStandard()
     {
         $_SERVER['SERVER_SOFTWARE'] = 'Google App Engine';
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
         $this->assertInstanceOf(
             'Google\Auth\Credentials\AppIdentityCredentials',
             ApplicationDefaultCredentials::getCredentials()
@@ -732,7 +732,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $_SERVER['SERVER_SOFTWARE'] = 'Google App Engine';
         putenv('GAE_INSTANCE=aef-default-20180313t154438');
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
         $httpHandler = getHandler([
             new Response(200, [GCECredentials::FLAVOR_HEADER => 'Google']),
         ]);
@@ -746,7 +746,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     {
         $_SERVER['SERVER_SOFTWARE'] = 'Google App Engine';
         putenv('GAE_INSTANCE=aef-default-20180313t154438');
-        putenv('HOME=' . __DIR__ . '/not_exist_fixtures');
+        setHomeEnv(__DIR__ . '/not_exist_fixtures');
         $httpHandler = getHandler([
             new Response(200, [GCECredentials::FLAVOR_HEADER => 'Google']),
         ]);
@@ -868,7 +868,7 @@ class ApplicationDefaultCredentialsTest extends TestCase
     /** @runInSeparateProcess */
     public function testUniverseDomainInGceCredentials()
     {
-        putenv('HOME');
+        setHomeEnv(null);
 
         $expectedUniverseDomain = 'example-universe.com';
         $creds = ApplicationDefaultCredentials::getCredentials(